+++ packages/api/src/app.module.ts
@@ -46,10 +46,10 @@ import { LoggerModule } from './logger/logger.module';
 import { MailerModule } from './mailer/mailer.module';
 import { MigrationModule } from './migration/migration.module';
 import { SettingModule } from './setting/setting.module';
+import { WorkflowTransferModule } from './transfer/workflow-transfer.module';
 import { Ability } from './user/guards/ability.guard';
 import { UserModule } from './user/user.module';
 import { WebsocketModule } from './websocket/websocket.module';
-import { WorkflowTransferModule } from './workflow/transfer/workflow-transfer.module';
 import { WorkflowModule } from './workflow/workflow.module';
 
 // Production "monolith" mode
