+++ packages/api/src/transfer/index.ts
@@ -1,0 +1,19 @@
+/*
+ * Hexabot — Fair Core License (FCL-1.0-ALv2)
+ * Copyright (c) 2026 Hexastack.
+ * Full terms: see LICENSE.md.
+ */
+
+export * from './workflow-transfer.module';
+
+export * from './workflow-transfer.controller';
+
+export * from './workflow-transfer-definition.service';
+
+export * from './workflow-transfer-resource-adapter';
+
+export * from './workflow-transfer-adapter.registry';
+
+export * from './workflow-transfer.types';
+
+export * from './workflow-transfer.service';
