+++ packages/api/src/workflow/index.ts
@@ -72,18 +72,18 @@ export * from './services/mcp-server.service';
 
 export * from './services/mcp-client-pool.service';
 
-export * from './transfer/workflow-transfer.module';
+export * from '../transfer/workflow-transfer.module';
 
-export * from './transfer/workflow-transfer.controller';
+export * from '../transfer/workflow-transfer.controller';
 
-export * from './transfer/workflow-transfer-definition.service';
+export * from '../transfer/workflow-transfer-definition.service';
 
-export * from './transfer/workflow-transfer-resource-adapter';
+export * from '../transfer/workflow-transfer-resource-adapter';
 
-export * from './transfer/workflow-transfer-adapter.registry';
+export * from '../transfer/workflow-transfer-adapter.registry';
 
-export * from './transfer/workflow-transfer.types';
+export * from '../transfer/workflow-transfer.types';
 
-export * from './transfer/workflow-transfer.service';
+export * from '../transfer/workflow-transfer.service';
 
 export * from './defaults/default-workflow';
